Estou tentando fazer a reserva baseada no outro branch, porém com algumas coisas diferentes

+++ src/controller/comando/PagarMulta.java
@@ -20,15 +20,20 @@ public class PagarMulta implements Command{
         Usuario usuario = udao.getUsuario(entrada.nextInt());
         entrada.nextLine();
 
-        for(Multa m : usuario.getMultas()){
-            System.out.println(m);
+        if(!(usuario.getMultas().isEmpty())){
+            for(Multa m : usuario.getMultas()){
+                System.out.println(m);
+            }
+
+            System.out.println("Entre com o id da multa a pagar:");
+            mdao.remove(entrada.nextInt());
+            entrada.nextLine();
+
+            System.out.println("Multa paga com sucesso!");
+        }else {
+            System.out.println("Não há multas a pagar!");
         }
 
-        System.out.println("Entre com o id da multa a pagar:");
-        mdao.remove(entrada.nextInt());
-        entrada.nextLine();
-
-        System.out.println("Multa paga com sucesso!");
         entrada.nextLine();
     }
 }
